did some flaveicons tuff
